requirements.txt
alembic==0.8.9
click==6.6
dominate==2.3.1
enum==0.4.6
Flask==0.11.1
Flask-Bootstrap==3.3.7.0
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.6
MarkupSafe==0.23
Pillow==3.4.2
python-editor==1.0.3
pytz==2016.10
SQLAlchemy==1.1.4
visitor==0.1.3
Werkzeug==0.11.11
